Add DB class, delete kim system

+++ app/models/Kim.php
@@ -4,13 +4,18 @@ include '../app/Model.php';
 
 class Kim extends Model {
 	
+	private $kimsDB;
+
 	public function __construct() {
-		return parent::__construct('Kims');
+		$this->kimsDB = new DB('Kims');
 	}
 
 	public function add($post,$files) {
 		$uploaded = $files['zip']['error']==UPLOAD_ERR_OK;
 		$errors = [];
+		if(!$uploaded) {
+			array_push($errors,'File not uploaded');
+		}
 		if(empty($post)) {
 			array_push($errors,'Отсутствует номер КИМа');
 			$kim_number = -1;
@@ -22,14 +27,20 @@ class Kim extends Model {
 		$name = $files['zip']['tmp_name'];
 		$tmp_path = '../tempfiles';
 		$zip_path = $tmp_path.'/kim'.$kim_number.'.zip';
-		move_uploaded_file($name,$zip_path);
-		$zip = new ZipArchive;
-		$zip->open($zip_path);
-		$zip->extractTo($tmp_path);
-		$zip->close();
-		unlink($zip_path);
+		$moved = move_uploaded_file($name,$zip_path);
+		if($moved) {
+			$zip = new ZipArchive;
+			$opened = $zip->open($zip_path);
+			if($opened===true) {
+				$zip->extractTo($tmp_path);
+				$zip->close();
+			}
+			unlink($zip_path);
+		}
+		$kim_data = scandir($tmp_path);
+		$kim_data = array_diff($kim_data,['.','..']);
 		if(!file_exists($tmp_path.'/ans.json')) {
-			array_push($errors,'No file "ans.json"');
+			array_push($errors,'No file ans.json');
 			$kim_ans = [];
 		}
 		else {
@@ -40,46 +51,57 @@ class Kim extends Model {
 		for($i=1; array_key_exists($i,$kim_ans); $i++) {
 			$task_count++;
 		}
+		for($i=1; $i<=$task_count; $i++) {
+			if(!in_array($i.'.png',$kim_data)) {
+				array_push($errors,'No file '.$i.'.png');
+			}
+		}
+		if(!in_array('info.png',$kim_data)) {
+			array_push($errors,'No file info.png');
+		}
+		$kim_files = [];
 		for($i=1;$i<=$task_count;$i++) {
 			$kim_files[$i] = md5('kim'.$kim_number.'number'.$i).'.png';
 		}
-		if(file_exists($tmp_path.'/info.png')) {
-			$kim_files['i'] = md5('kim'.$kim_number.'info').'.png';
-		}
-		else {
-			array_push($errors,'No file "info.png"');
-		}
+		$kim_files['i'] = md5('kim'.$kim_number.'info').'.png';
 		$kim_info = [
 			'task_count' => $task_count,
 			'ans' => $kim_ans,
 			'files' => $kim_files
 		];
-		for($i=1;$i<=$task_count;$i++) {
-			if(!copy($tmp_path.'/'.$i.'.png','img/'.$kim_files[$i])) {
-				array_push($errors,'No file "'.$i.'.png"');
-			}
-			unlink($tmp_path.'/'.$i.'.png');
-		}
-		copy($tmp_path.'/info.png','img/'.$kim_files['i']);
-		unlink($tmp_path.'/info.png');
-		unlink($tmp_path.'/ans.json');
-		if(empty($errors)) {
-			$this->db_set($kim_number,$kim_info);
-			$msg = 'Успешно';
+		if(count($errors)) {
+			$join = implode('\n',$errors);
+			$msg = 'Ошибки:\n'.$join;
 		}
 		else {
-			$msg = 'Ошибки:\n'.implode('\n',$errors);
+			for($i=1;$i<=$task_count;$i++) {
+				copy($tmp_path.'/'.$i.'.png','img/'.$kim_files[$i]);
+			}
+			copy($tmp_path.'/info.png','img/'.$kim_files['i']);
+			$this->kimsDB->set($kim_number,$kim_info);
+			$msg = 'Успешно';
+		}
+		if(count($kim_data)) {
+			foreach($kim_data as $file) {
+				unlink($tmp_path.'/'.$file);
+			}
 		}
 		return $msg;
 	}
 
 	public function delete($post) {
-		$kim = $this->db_get($post['kim']);
-		$this->db_unset($post['kim']);
-		for($i=1;$i<=$kim['task_count'];$i++) {
-			unlink('img/'.$kim['files'][$i]);
+		$kims = json_decode($post['json'],false);
+		foreach($kims as $kim) {
+			$files = $this->kimsDB->get($kim)['files'];
+			foreach($files as $file) {
+				unlink('img/'.$file);
+			}
+			$this->kimsDB->unset($kim);
 		}
-		unlink('img/'.$kim['files']['i']);
+	}
+
+	public function getKims() {
+		return $this->kimsDB->keys();
 	}
 
 	public function sample() {
